% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN.R
\name{BIEN_occurrence_spatialpolygons}
\alias{BIEN_occurrence_spatialpolygons}
\title{Extract occurrence data for specified SpatialPolygons or SpatialPolygonsDataFrame}
\usage{
BIEN_occurrence_spatialpolygons(
  spatialpolygons,
  cultivated = FALSE,
  only.new.world = FALSE,
  all.taxonomy = FALSE,
  native.status = FALSE,
  natives.only = TRUE,
  observation.type = FALSE,
  political.boundaries = FALSE,
  collection.info = F,
  ...
)
}
\arguments{
\item{spatialpolygons}{An object of class SpatialPolygons or SpatialPolygonsDataFrame.  Note that the file must be in WGS84.}

\item{cultivated}{Return known cultivated records as well?  Default is FALSE.}

\item{only.new.world}{Return only records from the New World?  Default is FALSE}

\item{all.taxonomy}{Return all taxonomic information?  This includes the raw data as well as the "scrubbed" data.}

\item{native.status}{Return information on introduction status?  The default value is FALSE. A value of TRUE also returns additional information on introduction status.}

\item{natives.only}{Exclude detected introduced species?  Default is TRUE.}

\item{observation.type}{Return information on type of observation (i.e. specimen vs. plot)?  The default value is FALSE.}

\item{political.boundaries}{Return information on political boundaries for an observation? The default value is FALSE.}

\item{collection.info}{Return additional information about collection and identification? The default value is FALSE.}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
Dataframe containing occurrence records for the specified species.
}
\description{
BIEN_occurrence_spatialpolygons downloads occurrence records falling within a user-specified SpatialPolygons or SpatialPolygonsDataFrame.
}
\note{
US FIA coordinates have been fuzzed and swapped, for more details see: https://www.fia.fs.fed.us/tools-data/spatial/Policy/index.php

We recommend using \code{\link[rgdal]{readOGR}} to load spatial data
}
\examples{
\dontrun{
library(rgdal)
BIEN_ranges_species("Carnegiea gigantea")#saves ranges to the current working directory
sp<-readOGR(dsn = ".",layer = "Carnegiea_gigantea")
#SpatialPolygons should be read with readOGR().
species_occurrences<-BIEN_occurrence_spatialpolygons(spatialpolygons=sp)}
}
\seealso{
Other occurrence functions: 
\code{\link{BIEN_occurrence_box}()},
\code{\link{BIEN_occurrence_country}()},
\code{\link{BIEN_occurrence_county}()},
\code{\link{BIEN_occurrence_family}()},
\code{\link{BIEN_occurrence_genus}()},
\code{\link{BIEN_occurrence_records_per_species}()},
\code{\link{BIEN_occurrence_species}()},
\code{\link{BIEN_occurrence_state}()}
}
\concept{occurrence functions}
