% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN.R
\name{BIEN_phylogeny_conservative}
\alias{BIEN_phylogeny_conservative}
\title{Download the conservative BIEN phylogeny}
\usage{
BIEN_phylogeny_conservative(...)
}
\arguments{
\item{...}{Additional arguments passed to .BIEN_sql}
}
\value{
A phylo object containing the BIEN conservative phylogeny
}
\description{
BIEN_phylogeny_conservative downloads the conservative BIEN phylogeny, which only includes species with molecular data available.
}
\examples{
\dontrun{
BIEN_phylo<-BIEN_phylogeny_conservative()}
}
\seealso{
Other phylogeny functions: \code{\link{BIEN_phylogeny_complete}}
}

