% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN.R
\name{BIEN_plot_dataset}
\alias{BIEN_plot_dataset}
\title{Download plot data by dataset.}
\usage{
BIEN_plot_dataset(dataset, cultivated = FALSE, only.new.world = TRUE,
  all.taxonomy = FALSE, print.query = FALSE, native.status = FALSE,
  political.boundaries = FALSE, all.metadata = FALSE, ...)
}
\arguments{
\item{dataset}{A plot dataset or vector of datasets. See BIEN_plot_metadata for more information on plots.}

\item{cultivated}{Return cultivated records as well?  Default is FALSE.}

\item{only.new.world}{Return only records from the New World?  Default is true}

\item{all.taxonomy}{Return all taxonomic information?  This includes the raw data as well as the "scrubbed" data.}

\item{print.query}{Should the PostgreSQL query be printed? The default value is FALSE.}

\item{native.status}{Return information on introduction status?  The default value is FALSE. A value of TRUE also returns additional information on introduction status.}

\item{political.boundaries}{Return information on political boundaries for an observation? The default value is FALSE.}

\item{all.metadata}{Should additional plot metadata be returned?  Default is FALSE.}

\item{...}{Additional arguments passed to .BIEN_sql}
}
\value{
A dataframe containing all data from the specified dataset.
}
\description{
BIEN_plot_dataset downloads all plot data for a given dataset or datasets.
}
\examples{
\dontrun{
BIEN_plot_dataset("Gentry Transect Dataset")}
}
\seealso{
Other plot functions: \code{\link{BIEN_plot_country}},
  \code{\link{BIEN_plot_datasource}},
  \code{\link{BIEN_plot_list_datasource}},
  \code{\link{BIEN_plot_list_sampling_protocols}},
  \code{\link{BIEN_plot_metadata}},
  \code{\link{BIEN_plot_name}},
  \code{\link{BIEN_plot_sampling_protocol}},
  \code{\link{BIEN_plot_state}}
}

