% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN.R
\name{BIEN_ranges_species}
\alias{BIEN_ranges_species}
\title{Download range maps for given species.}
\usage{
BIEN_ranges_species(species, directory = NULL, matched = TRUE,
  match_names_only = FALSE, include.gid = FALSE, print.query = FALSE, ...)
}
\arguments{
\item{species}{A single species or a vector of species.}

\item{directory}{Directory that range maps should be saved in.  If none is specified, range maps will be saved in the current working directory.}

\item{matched}{Return a list of species that were downloaded. Default is TRUE.}

\item{match_names_only}{Check for range maps for the species specified without downloading range maps. Default is FALSE.}

\item{include.gid}{Should the files returned have a unique GID appended to them? This is needed if downloading multiple maps for the same species.}

\item{print.query}{Should the PostgreSQL query be printed? The default value is FALSE.}

\item{...}{Additional arguments passed to BIEN_sql}
}
\value{
Range maps for specified species.
}
\description{
BIEN_ranges_species extracts range maps for the specified species.
}
\examples{
\dontrun{
species_vector<-c("Abies_lasiocarpa","Abies_amabilis")
testwd<-"C:/wherever/you/want/files/saved/" #Set a working directory
BIEN_ranges_species(species_vector)
BIEN_ranges_species(species_vector,match_names_only = TRUE)
BIEN_ranges_species(species_vector,test_wd)#saves ranges to a specified working directory
BIEN_ranges_species("Abies_lasiocarpa")
BIEN_ranges_species("Abies_lasiocarpa","C:/wherever/you/want/files/saved/")

#Reading files
setwd("C:/wherever/your/shapefiles/are/")
Abies_poly<-readShapePoly("Abies_lasiocarpa")
Abies_poly<-readShapePoly("C:/wherever/your/shapefiles/are/Abies_lasiocarpa.shp")

#Plotting files
plot(Abies_poly)#plots the shapefile, but doesn't mean much without any reference
require(maps)#easy source of maps
map('world', fill = TRUE, col = "grey")#plots a world map (WGS84 projection), in grey
plot(Abies_poly,col="forest green",add=TRUE) #adds the range of Abies lasiocarpa to the map

#Getting data from the files (currently only species names)
Abies_poly$Species#gives the species name associated with "Abies_poly"}
}
\seealso{
Other range functions: \code{\link{BIEN_ranges_box}},
  \code{\link{BIEN_ranges_genus}},
  \code{\link{BIEN_ranges_intersect_species}},
  \code{\link{BIEN_ranges_load_species}},
  \code{\link{BIEN_ranges_shapefile}}
}

