% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN.R
\name{BIEN_stem_genus}
\alias{BIEN_stem_genus}
\title{Extract stem data for specified genera from BIEN}
\usage{
BIEN_stem_genus(genus, cultivated = FALSE, only.new.world = TRUE,
  all.taxonomy = FALSE, print.query = FALSE, native.status = FALSE,
  political.boundaries = FALSE, all.metadata = F, ...)
}
\arguments{
\item{genus}{A single genus, or a vector of genera. Genera should be capitalized.}

\item{cultivated}{Return known cultivated records as well?  Default is FALSE.}

\item{only.new.world}{Return only records from the New World?  Default is true}

\item{all.taxonomy}{Return all taxonomic information?  This includes the raw data as well as the "scrubbed" data.}

\item{print.query}{Should the PostgreSQL query be printed? The default value is FALSE.}

\item{native.status}{Return information on introduction status?  The default value is FALSE. A value of TRUE also returns additional information on introduction status.}

\item{political.boundaries}{Return information on political boundaries for an observation? The default value is FALSE.}

\item{all.metadata}{Should additional plot metadata be returned?  Default is FALSE.}

\item{...}{Additional arguments passed to BIEN_sql}
}
\value{
Dataframe containing stem data for the specified genera.
}
\description{
BIEN_stem_genus downloads occurrence records for specific genera from the BIEN database.
}
\note{
Setting either "cultivated" or "native.status" to TRUE will significantly slow the speed of a query.
}
\examples{
\dontrun{
BIEN_stem_genus(genus = "Tovomita")
genus_vector<-c("Tovomita", "Myrcia")
BIEN_stem_genus(genus = genus_vector)
BIEN_stem_genus(genus = genus_vector, all.taxonomy=TRUE)}
}
\seealso{
Other stem functions: \code{\link{BIEN_stem_family}},
  \code{\link{BIEN_stem_species}}
}

