% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN.R
\name{BIEN_taxonomy_family}
\alias{BIEN_taxonomy_family}
\title{Extract taxonomic information for families}
\usage{
BIEN_taxonomy_family(family, print.query = FALSE, ...)
}
\arguments{
\item{family}{A single family or a vector of families.}

\item{print.query}{Should the PostgreSQL query be printed? The default value is FALSE.}

\item{...}{Additional arguments passed to .BIEN_sql}
}
\value{
Dataframe containing taxonomic information for the specified families.
}
\description{
BIEN_taxonomy_family downloads a dataframe of all taxonomic information for given families.
}
\examples{
\dontrun{
BIEN_taxonomy_family("Orchidaceae")
family_vector<-c("Orchidaceae","Poaceae")
BIEN_taxonomy_family(family_vector)}
}
\seealso{
Other taxonomy functions: \code{\link{BIEN_taxonomy_genus}},
  \code{\link{BIEN_taxonomy_species}}
}

