% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN.R
\name{BIEN_trait_genus}
\alias{BIEN_trait_genus}
\title{Download trait data for given genera.}
\usage{
BIEN_trait_genus(genus, ...)
}
\arguments{
\item{genus}{A single genus or a vector of genera.}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
A dataframe of all data matching the specified genera.
}
\description{
BIEN_trait_genus extracts entries that contain the specified genera.
}
\examples{
\dontrun{
BIEN_trait_genus("Acer")
genus_vector<-c("Acer","Abies")
BIEN_trait_genus(genus_vector)}
}

