% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.collection_check_plot}
\alias{.collection_check_plot}
\title{Set query details}
\usage{
.collection_check_plot(collection.info)
}
\description{
Helper function to set query components.
}
\keyword{internal}
