% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.genus_check}
\alias{.genus_check}
\title{Set query details}
\usage{
.genus_check(genus)
}
\arguments{
\item{genus}{Single genus or vector of genera.}
}
\description{
Helper function to set query components.
}
\keyword{internal}
