% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.taxonomy_check_traits}
\alias{.taxonomy_check_traits}
\title{Set query details}
\usage{
.taxonomy_check_traits(all.taxonomy)
}
\description{
Helper function to set query components.
}
\keyword{internal}
